+++ provider.tf
@@ -1,5 +1,0 @@
-provider "aws" {
-  region = var.aws_region
-  #version = "~> 1.54"
-}
-